Simplify source-remove handling
Before the change we were constructing lines to be removed based on SSRC
parameters values supplied in the notification. We want the client to simply
remove all lines related to the SSRC being removed as sometimes things get
out of sync. For example when Chrome adds 'default' mslabel or when we'll be
doing lipsync stream merging.