Changes the way we initialize callstats.
Call addNewFabric only if callstats is initialized, there are certain situations when the init callback for some reason is slow and we call addNewFabric before callstats is initialized and we skip the peerconnection reported so we are missing the stats. Reports data only if callstats had initialized, adds caching of associateMstWithUserID calls.