Be forgiving towards API clients
If the API client happened to try to add one and the same local video
track twice, the library would throw an error. I find that draconian
because:

1. The error message states that a second local video stream cannot be
added but that is incorrect because the API client is not trying to add
a second local video stream but add one and the same local video stream
multiple times.

2. The jsdoc says that a throw will happen if another local video stream
is added but that is misleading because the API client is not trying to
add another local video stream but add the same local video stream.

3. Adding one and the same local video stream multiple times can be
handled gracefully by the lib-jitsi-meet library. Thus, the library
appears more friendly.

Hence, gracefully handle the case of adding the same local video track
multiple times without throwing an error.