Revert Promise reject reason
I added reporting of Promise's reject reason in b10a1fc while merging
jitsi/ss_postinstall_fix since Promise's reject function has such an
argument. Hristo Terezov disagreed with it because reason is undefined
in the use case and found it to be confusing so I'm reverting to what he
intended in the first place.