ref: introduce 'ModificationQueueTask'
The purpose of ModificationQueueTask is to make it clearer on
what are the possible outcomes of the worker function execution
and they way the results can be set by the worker function. Now
they are passed through the task instance using appropriate setters
in contrary to the possibility to return few different types of values.