Adds peer connection statuses (active/inactive/interrupted/restoring). (#430)
* Adds peer connection statuses (active/inactive/interrupted/restoring).

The peer connection used to be check with is connection active. Adding new states like active, inactive, when the connection was intentionally stopped by the bridge, interrupted due to network problem and restoring, which is the state of a peer which was inactive as being out of lastN and is now entering lastM and eventually will become active.

* Adds timers to track restoring status.

If restoring too long we set the connection status to interrupted. We save timestamps for all participants entering lastN and set the status for the connection to restoring for 5 seconds, or till it goes into active or exits lastN.

* Adds initial connectionStatus of newly created participant - active.

* Fixes some jsdocs.

* Uses symbols for strings and use of map, fixing comments on PR.

* Removes symbols.