fix(logs&stats): BEFORE_STATISTICS_DISPOSED
Will emit BEFORE_STATISTICS_DISPOSED event, before the last CallStats
fabric is terminated to make sure that the final logs batch is reported
correctly.

Also adds a check for CallStats instance to queue events when
the backend has been initialized, but there is no CallStats instance
available to report on.