fix(ParticipantConnectionStatus): p2p and lastN
Splits the participant connection status detection logic into two
separate flows one for P2P and one for JVB.

Fixes problem where user outside of last N would be reported as INACTIVE
when in P2P mode.

Added some improvements to the detection logic by adding more priority
for the video tracks frozen detection (if supported) over the last N.
When the browser supports this detection and it's playing video we
should not check the last N, because it's less reliable.