fix(permission-prompt): Use labels to predict if the prompt will not display
In Firefox, getUserMedia for audio and video can take longer than
USER_MEDIA_PERMISSION_PROMPT_TIMEOUT, causing an event to fire stating the
permission prompt is displayed even when it is not. On my machine,
getUserMedia for audio and video would take between 200ms and 700ms. Due to
the variance, relying on the timeout alone to know about permission prompt
display might not be sufficient. One way to know if permission has been
granted, thereby knowing the prompt will not display, is by executing
enumerateDevices and seeing if devices have labels. The call to enumerate
is consistently and significantly faster that getUserMedia. This will
not work on temasys though because enumerateDevices will always return
labels.