Last n improvements (#493)
* doc(JitsiConference): deprecate 'isInLastN'

The 'isInLastN' method should not be used for the UI purposes, but
ParticipantConnectionStatus value should be used instead.

* fix(ParticipantConn..Status): speed up INACTIVE transition

Before this change when user's video stops playing, after user is
removed from last N we were waiting 2 seconds, before going to INACTIVE
state. This commit reduces the time to 500ms for such case.

* fix(ParticipantConn...): reduce logging

Reduce logging verbosity.

* fix(ParticipantConn...): handle LastN == 0

When LastN is set to 0 we should not rely on video playback and last N
set for figuring out participant connection status.

* fix(JitsiConference): undefined participants

Fixes a crash when this.participants field is accessed from _init.