fix(permission-prompt): increase timeout for displaying permission prompt
Firefox's getUserMedia can take longer than 500ms to complete. This
causes the permission prompt timeout to fire even if the user has
given permission to audio and video. There may not be a reliable
way right now to check if the user has given permission to both
audio and video so dumbly increase the timeout instead.