P2P video track added later (#515)
* ref(JSPC): simplify SSRC owner in P2P

There's no need for any extra extensions for SSRCs owner signalling in
P2P, because it's always the remote peer who owns them.

This also fixes a problem where no SSRC owner was added for 'source-add'
in P2P (for JVB conference Jicofo adds that).

* fix(TPC): always advertise 'sendrecv'

Our media direction is only ever updated on the remote side with
the initial offer (or answer). Because of that we want to advertise
'sendrecv' even if we start with no video (or audio) track.

It is OK to adjust this direction in the localDescription getter,
because it's adjusted again in the setter to the correct value based on
local tracks, so the SDP transformation chain still works fine.