ref: Refactors average statistics reporting.
This commit changes the way the average stats are being reported to the
analytics backend. Instead of prefixes for the different scenarios
("p2p_", "jvb121_" and no prefix for multi-party calls) we now have
properties that allow to distinguish which scenario an event pertains
to. The query language of the analytics backend can be used to build
appropriate queries that group by those properties.

Additionally, we now report whether the average stats event concerns a
relayed call (a call that's routed via a TURN server) or not.