fix(JitsiConference): reduce verbosity level
There are scenarios when it's ok to call setP2PStatus again with
the same value. For example when P2P is stopped, before it starts with
the A/B testing mode enabled. It's only important to know that it
happened, but it's not an error, because the code will not execute
and return immediately.