[RN] Fix error on undefined variable
adapter.screenshare.js uses the variable requestUserMedia without
defining it. And that's a problem because it's packed in strict mode so
it will cause an error to be thrown at runtime which will fail
lib-jitsi-meet and, consequently, the whole app. As the variable
requestUserMedia appears to be global in certain browser, prevent the
use of the undefined variable without really changing the outcome of its
use.