API.md - Important Mention of Add Local Track to Conference
In most use cases, the developer will want to add the newly acquired local track to the room.
As this is extremely common and required to actually share the track with others I would like to recommend
that step 5 mentions this. This makes it clear that createLocalTracks does not automatically do anything like this.