ref: delay ICE failed event
Adds extra 15 seconds before emitting ICE failed event. If the OS is capable of
reporting internet offline status then will start counting when the internet
comes back online.

The reason for that is when ICE restarts are disabled a user will get the page reload
screen immediately, but there's a chance to recover as long as the channels have not
expired on the bridge.