Fix of case when findLine returns false
Hi,

I stumbled upon a case where `findLine` at this position in the code returns `false` instead of a string (I'm not entirely sure what the case is about regarding the actual feature though, I arrived there through the debugger). The problem is that it makes the .substring(7) fail because the function does not exist on `false`. The proposed change fixes this case.