feat(perf-metrics) introduce a module to take performance metrics
The perf-metrics module is a thin wrapper around the W3C Performance API and
PerformanceObserver which makes it easy to take measurements of internal code
operations.

The API is simple, 2 methods are enough to get started: markStart and markEnd.
These are shortcuts over the performance "measures" API, but the performance API
can be used directly just fine.

As an example let's look at the XMPP MUC join time. markStart is called
when the join() function is called and markEnd is called when the operation has
completed. The resulting metric is logged with its duration.

Metric names need not be unique.

The perf-metrics framework is exposed in the JitsiMeetJS object for embedding
applications to use. So long as the metrics begin with the name "jitsi." they
will be logged by the observer.

This replaces the previous getConnectionTimes() custom implementation.