ref: add getSourceIndexFromSourceName
Adds a helper function getSourceIndexFromSourceName to
have track index parsing in one place.

In _processLocalSSRCsMap there was an assumption that a track
index can be a single digit only.

Use "last index of" or search from the end of the collection to
make parsing a bit more universal (does not matter as much for
Jitsi endpoint IDs, but if fixes issues if someone has
a customization there).