fix: Only retry on "invalid sessin" errors
We used to fire FOCUS_DISCONNECTED (which doesn't have anything to do
with anything disconnecting) and then retrying. This is not necessary
because FOCUS_DISCONNECTED is a fatal error and the client schedules
a reload of it's own (and we want the client reload as opposed to the
conference-request retry in case e.g. the shard has changed).